make some fixes
