+++ frontend/src/APIClients/queries/StoryQueries.ts
@@ -115,47 +115,3 @@ export const buildHomePageStoriesQuery = (
 
   return result as QueryInformation;
 };
-
-const COMMENT_FIELDS = `
-    id
-    commentIndex
-    content
-    storyTranslationContentId
-    `;
-
-export const buildCommentsQuery = (
-  storyTranslationId: number,
-  resolved: boolean | null,
-): QueryInformation => {
-  let queryString = gql`
-    query CommentsByStoryTranslation {
-      commentsByStoryTranslation( storyTranslationId: ${storyTranslationId} ){
-        ${COMMENT_FIELDS}
-      }
-    }
-  `;
-
-  if (resolved === true || resolved === false) {
-    queryString = gql`
-    query CommentsByStoryTranslation {
-      commentsByStoryTranslation( storyTranslationId: ${storyTranslationId}, resolved: ${resolved} ){
-        ${COMMENT_FIELDS}
-      }
-    }
-  `;
-  }
-
-  return {
-    fieldName: "commentsByStoryTranslation",
-    string: queryString,
-  } as QueryInformation;
-};
-
-export type CommentResponse = {
-  id: number;
-  userId: number;
-  commentIndex: number;
-  time: string;
-  resolved: boolean;
-  content: string;
-};
