--- frontend/src/assets/Redo.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Icon } from "@chakra-ui/icon";
-
-const RedoIcon = () => {
-  return (
-    <Icon>
-      <path
-        d="M13.8462 6.15385H6.46154C4.42585 6.15385 2.76923 7.81046 2.76923 9.84615C2.76923 11.8818 4.42585 13.5385 6.46154 13.5385H10.1538V16H6.46154C3.06831 16 0.307692 13.2394 0.307692 9.84615C0.307692 6.45292 3.06831 3.69231 6.46154 3.69231H13.8462V0L20 4.92308L13.8462 9.84615V6.15385Z"
-        fill="#606062"
-      />
-    </Icon>
-  );
-};
-
-export default RedoIcon;
