--- frontend/src/assets/Undo.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Icon } from "@chakra-ui/icon";
-
-const UndoIcon = () => {
-  return (
-    <Icon>
-      <path
-        d="M6.15385 6.15385H13.5385C15.5742 6.15385 17.2308 7.81046 17.2308 9.84615C17.2308 11.8818 15.5742 13.5385 13.5385 13.5385H9.84615V16H13.5385C16.9317 16 19.6923 13.2394 19.6923 9.84615C19.6923 6.45292 16.9317 3.69231 13.5385 3.69231H6.15385V0L0 4.92308L6.15385 9.84615V6.15385Z"
-        fill="#606062"
-      />
-    </Icon>
-  );
-};
-
-export default UndoIcon;
